+++ Changelog.md
@@ -3,6 +3,7 @@
 ## Version (NEXT)
 * added heavy-hex-topology (up to 127 qubits)
 * added QPU data to backend namespace
+* added basis-strings and basis-labels functions
 * refactored error mitigation
 * updated tutorial